requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
apscheduler
python-dotenv
pydantic
langchain-google-genai
langchain-community
autogen
dateparser